Move client into separate file
