+++ crates/communication/src/server/parameters/mod.rs
@@ -2,20 +2,20 @@ use serde_json::Value;
 
 use crate::messages::{ParametersRequest, Path};
 
-use super::Client;
+use super::client::Client;
 
 pub mod directory;
 pub mod storage;
 pub mod subscriptions;
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ClientRequest {
+pub struct ClientRequest {
     pub request: ParametersRequest,
     pub client: Client,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) enum StorageRequest {
+pub enum StorageRequest {
     UpdateParameter {
         client: Client,
         id: usize,
